reorg docs
