+++ docs/app/(content)/layout.module.css
