+++ docs/app/(content)/react/conditional/page.mdx
@@ -1,4 +1,4 @@
-import { DemoSection } from '../../components/DemoSection';
+import { DemoSection } from '../../../../components/DemoSection';
 import * as conditionalDemo from './demo/Conditional/config?demo';
 
 export const metadata = {
