+++ docs/app/(content)/react/custom-resizer/demo/ResizerStyle01/index.tsx
