+++ docs/app/(content)/react/custom-resizer/demo/ResizerStyle02/index.css
