+++ docs/app/(content)/react/custom-resizer/demo/ResizerStyle03/index.tsx
