+++ docs/app/(content)/react/custom-resizer/demo/ResizerStyle04/index.css
