+++ docs/app/(content)/react/custom-resizer/page.mdx
@@ -1,4 +1,4 @@
-import { DemoSection } from '../../components/DemoSection';
+import { DemoSection } from '../../../../components/DemoSection';
 import * as resizerStyle01Demo from './demo/ResizerStyle01/config?demo';
 import * as resizerStyle02Demo from './demo/ResizerStyle02/config?demo';
 import * as resizerStyle03Demo from './demo/ResizerStyle03/config?demo';
