+++ docs/app/(content)/react/nested/page.mdx
@@ -1,4 +1,4 @@
-import { DemoSection } from '../../components/DemoSection';
+import { DemoSection } from '../../../../components/DemoSection';
 import * as nestedDemo from './demo/NestedDemo/config?demo';
 
 export const metadata = {
