+++ docs/app/(content)/react/page.mdx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
-import { DemoSection } from '../components/DemoSection';
-import { BundleJsBadge } from '../components/BundleJsBadge';
+import { DemoSection } from '../../../components/DemoSection';
+import { BundleJsBadge } from '../../../components/BundleJsBadge';
 import * as nestedDemo from './nested/demo/NestedDemo/config?demo';
 
 export const metadata = {
