+++ docs/app/(content)/react/persistence/page.mdx
@@ -1,4 +1,4 @@
-import { DemoSection } from '../../components/DemoSection';
+import { DemoSection } from '../../../../components/DemoSection';
 import * as persistenceDemo from './demo/PersistenceDemo/config?demo';
 
 export const metadata = {
