+++ docs/app/(content)/react/resizable/demo/HorizontalResizableDemo/config.ts
