+++ docs/app/(content)/react/resizable/demo/HorizontalResizableDemo/index.css
