+++ docs/app/(content)/react/resizable/demo/IframeDemo/config.ts
