+++ docs/app/(content)/react/resizable/demo/IframeDemo/index.css
