+++ docs/app/(content)/react/resizable/demo/VerticalResizableDemo/config.ts
