+++ docs/app/(content)/react/resizable/demo/VerticalResizableDemo/index.css
