+++ docs/app/(content)/react/resizable/page.mdx
@@ -1,4 +1,4 @@
-import { DemoSection } from '../../components/DemoSection';
+import { DemoSection } from '../../../../components/DemoSection';
 import * as horizontalDemo from './demo/HorizontalResizableDemo/config?demo';
 import * as verticalDemo from './demo/VerticalResizableDemo/config?demo';
 import * as iframeDemo from './demo/IframeDemo/config?demo';
