+++ docs/app/(content)/react/static/demo/ThreePaneDemo/config.ts
