+++ docs/app/(content)/react/static/demo/ThreePaneDemo/index.tsx
