+++ docs/app/(content)/react/static/demo/VerticalDemo/config.ts
