+++ docs/app/(content)/react/static/demo/VerticalDemo/index.tsx
