+++ docs/app/(content)/react/static/page.mdx
@@ -1,4 +1,4 @@
-import { DemoSection } from '../../components/DemoSection';
+import { DemoSection } from '../../../../components/DemoSection';
 import * as threePaneDemo from './demo/ThreePaneDemo/config?demo';
 import * as verticalDemo from './demo/VerticalDemo/config?demo';
