+++ docs/app/layout.tsx
@@ -1,34 +1,5 @@
-import '@/styles/globals.css';
-import '@split-ui/react/styles.css';
-import styles from './layout.module.css';
-import { Metadata } from 'next';
-import ResponsiveNav from '@/components/ResponsiveNav';
-import { SlugProvider } from '@/components/SlugContext';
+export const dynamic = 'error';
 
-export const metadata: Metadata = {
-  title: {
-    template: '%s | split-ui',
-    default: 'split-ui',
-  },
-  description:
-    'A React component library for creating resizable panel layouts with flexible sizing options and SSR support.',
-};
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html lang="en" data-scroll-behavior="smooth">
-      <body>
-        <div className={styles.layout}>
-          <ResponsiveNav />
-          <SlugProvider>
-            <main className={styles.main}>{children}</main>
-          </SlugProvider>
-        </div>
-      </body>
-    </html>
-  );
+export default function Layout({ children }: { children: React.ReactNode }) {
+  return children;
 }
